split IndMatch.hpp in two files for swig compatibility
