+++ src/aliceVision/matching/indMatch_test.cpp
@@ -5,7 +5,7 @@
 // v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "aliceVision/matching/IndMatch.hpp"
+#include "aliceVision/matching/MatchesCollections.hpp"
 #include "aliceVision/matching/io.hpp"
 
 #define BOOST_TEST_MODULE IndMatch
