+++ src/aliceVision/matching/kvld/algorithm.h
@@ -16,7 +16,7 @@
 
 #include <aliceVision/numeric/numeric.hpp>
 #include <aliceVision/image/Image.hpp>
-#include <aliceVision/matching/IndMatch.hpp>
+#include <aliceVision/matching/MatchesCollections.hpp>
 #include <aliceVision/feature/PointFeature.hpp>
 #include <aliceVision/types.hpp>
