+++ src/aliceVision/matchingImageCollection/geometricFilterUtils.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include <aliceVision/matching/IndMatch.hpp>
+#include <aliceVision/matching/MatchesCollections.hpp>
 #include <aliceVision/feature/FeaturesPerView.hpp>
 #include <aliceVision/camera/IntrinsicBase.hpp>
 #include <aliceVision/sfmData/SfMData.hpp>
